requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.40.0
openai>=1.0.0
Pillow>=10.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
undetected-chromedriver>=3.5.0
anthropic>=0.75.0
pydantic>=2.0.0
pytesseract>=0.3.10
schedule>=1.2.0
